+++ src/main/java/exceptions/LoaderMain.java
@@ -1,4 +1,4 @@
-package properties;
+package exceptions;
 
 import java.io.IOException;
 
@@ -6,7 +6,7 @@ public class LoaderMain {
     public static void main(String[] args) {
         try {
             Loader loader = new Loader("src/main/resources/file.properties");
-            String value = loader.getValue("dbUrl");
+            String value = loader.getValue("serverPort");
             System.out.println("Value from the file: " + value);
         } catch (IOException | IllegalArgumentException e) {
             System.out.println("Error " + e.getMessage());
