+++ src/main/java/exceptions/Main.java
@@ -1,4 +1,4 @@
-package properties;
+package exceptions;
 
 public class Main {
     public static void main(String[] args) {
